requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.2.0
Django==4.1.7
djangorestframework==3.14.0
flower==1.2.0
frozenlist==1.3.3
greenlet==2.0.1
humanize==4.6.0
idna==3.4
kombu==5.2.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.2
packaging==23.0
parsel==1.7.0
playwright==1.31.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pyee==9.0.4
python-decouple==3.8
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
yarl==1.8.2
